Master

+++ database/migrations/2020_11_23_000021_create_fuel_type_vehicle_pivot_table.php
@@ -9,9 +9,9 @@ class CreateFuelTypeVehiclePivotTable extends Migration
     public function up()
     {
         Schema::create('fuel_type_vehicle', function (Blueprint $table) {
-            $table->unsignedInteger('vehicle_id');
+            $table->unsignedBigInteger('vehicle_id');
             $table->foreign('vehicle_id', 'vehicle_id_fk_2499168')->references('id')->on('vehicles')->onDelete('cascade');
-            $table->unsignedInteger('fuel_type_id');
+            $table->unsignedBigInteger('fuel_type_id');
             $table->foreign('fuel_type_id', 'fuel_type_id_fk_2499168')->references('id')->on('fuel_types')->onDelete('cascade');
         });
     }
