+++ database/migrations/2020_11_23_000022_add_relationship_fields_to_employees_table.php
@@ -9,9 +9,9 @@ class AddRelationshipFieldsToEmployeesTable extends Migration
     public function up()
     {
         Schema::table('employees', function (Blueprint $table) {
-            $table->unsignedInteger('designation_id');
+            $table->unsignedBigInteger('designation_id');
             $table->foreign('designation_id', 'designation_fk_2587232')->references('id')->on('designations');
-            $table->unsignedInteger('office_id');
+            $table->unsignedBigInteger('office_id');
             $table->foreign('office_id', 'office_fk_2587233')->references('id')->on('offices');
         });
     }
