+++ database/migrations/2020_11_23_000023_add_relationship_fields_to_maintenance_histories_table.php
@@ -9,11 +9,11 @@ class AddRelationshipFieldsToMaintenanceHistoriesTable extends Migration
     public function up()
     {
         Schema::table('maintenance_histories', function (Blueprint $table) {
-            $table->unsignedInteger('region_name_id');
+            $table->unsignedBigInteger('region_name_id');
             $table->foreign('region_name_id', 'region_name_fk_2586335')->references('id')->on('regions');
-            $table->unsignedInteger('office_name_id');
+            $table->unsignedBigInteger('office_name_id');
             $table->foreign('office_name_id', 'office_name_fk_2586336')->references('id')->on('offices');
-            $table->unsignedInteger('registration_number_id');
+            $table->unsignedBigInteger('registration_number_id');
             $table->foreign('registration_number_id', 'registration_number_fk_2586337')->references('id')->on('vehicles');
         });
     }
