+++ database/migrations/2020_11_23_000024_add_relationship_fields_to_vehicles_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToVehiclesTable extends Migration
     public function up()
     {
         Schema::table('vehicles', function (Blueprint $table) {
-            $table->unsignedInteger('vehicle_type_id');
+            $table->unsignedBigInteger('vehicle_type_id');
             $table->foreign('vehicle_type_id', 'vehicle_type_fk_2499153')->references('id')->on('vehicle_types');
         });
     }
